refactor(integrity): shared IntegrityCheckResult→display selector
Tier 3 architecture review (A4).

The same IntegrityCheckResult was mapped to display data in two places:
IntegrityCheckCard (flattened detail lines) and SanctityDebugger
(structured diff panels), both off integritySync.state.result. Extracted
the derivation into src/utils/integrityDiff.ts, verbatim:
- hasInconsistencies(result)
- toDiffPanels(result)  — structured panels for the debugger
- yearDiffLines/awardDiffLines(diffs) — flattened lines for the card

Both components now consume the util. No behavior change. Suite 474
green, lint clean, build OK.
